chore: move getAttribute logic to utils
